// apps/web/src/api/generated/models/OpenClawCatalogChannelPlugin.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OpenClawCatalogChannelPlugin = {
    channelType?: string;
    displayName?: string;
    tier?: string;
    official?: boolean;
    connectionMode?: string;
    onboardingType?: string;
    enabledByPlatform?: boolean;
    requiredSecrets?: Array<string>;
    connectivityCheckMode?: string;
    messageTestMode?: string;
};